Cxx11 compat
